refactor: move `helix` into `pybes3` namespace
