--- docs/api/pybes3._cache_numba.md
@@ -1,12 +1,0 @@
-# pybes3._cache_numba
-
-!!! note "Internal API"
-    This module contains internal functions for managing numba caches. These are not part of the public API.
-
----
-::: pybes3._cache_numba.cache_auto_clear
----
-::: pybes3._cache_numba.check_numba_cache
----
-::: pybes3._cache_numba.clear_numba_cache
----
