+++ docs/api/pybes3.helix.md
@@ -1,0 +1,26 @@
+# pybes3.helix
+
+## Helix operations
+
+::: pybes3.helix.HelixObject
+---
+::: pybes3.helix.helix_obj
+---
+::: pybes3.helix.dr_phi0_to_x
+---
+::: pybes3.helix.dr_phi0_to_y
+---
+::: pybes3.helix.phi0_to_phi
+---
+::: pybes3.helix.kappa_to_pt
+---
+::: pybes3.helix.kappa_to_charge
+---
+::: pybes3.helix.kappa_to_radius
+---
+::: pybes3.helix.HelixAwkwardRecord
+---
+::: pybes3.helix.HelixAwkwardArray
+---
+::: pybes3.helix.helix_awk
+---
