--- docs/api/pybes3.tracks.md
@@ -1,26 +1,0 @@
-# pybes3.tracks
-
-## Helix operations
-
-::: pybes3.tracks.helix.HelixObject
----
-::: pybes3.tracks.helix.helix_obj
----
-::: pybes3.tracks.helix.dr_phi0_to_x
----
-::: pybes3.tracks.helix.dr_phi0_to_y
----
-::: pybes3.tracks.helix.phi0_to_phi
----
-::: pybes3.tracks.helix.kappa_to_pt
----
-::: pybes3.tracks.helix.kappa_to_charge
----
-::: pybes3.tracks.helix.kappa_to_radius
----
-::: pybes3.tracks.helix.HelixAwkwardRecord
----
-::: pybes3.tracks.helix.HelixAwkwardArray
----
-::: pybes3.tracks.helix.helix_awk
----
