+++ src/pybes3/besio/besio_cpp.pyi
@@ -3,17 +3,17 @@ from __future__ import annotations
 from typing import Any, Optional
 import numpy as np
 from numpy.typing import NDArray
-from uproot_custom.cpp import IElementReader
+from uproot_custom.cpp import IReader
 
-class Bes3TObjArrayReader(IElementReader):
+class Bes3TObjArrayReader(IReader):
     def __init__(
         self,
         name: str,
-        element_reader: IElementReader,
+        element_reader: IReader,
     ): ...
     def data(self) -> tuple[NDArray[np.uint32], Any]: ...
 
-class Bes3SymMatrixArrayReader(IElementReader):
+class Bes3SymMatrixArrayReader(IReader):
     def __init__(
         self,
         name: str,
@@ -22,14 +22,14 @@ class Bes3SymMatrixArrayReader(IElementReader):
     ): ...
     def data(self) -> NDArray[np.float64]: ...
 
-class Bes3CgemClusterColReader(IElementReader):
+class Bes3CgemClusterColReader(IReader):
     def __init__(self, name: str): ...
     def data(self) -> dict[str, NDArray]: ...
 
 def read_data(
     data: NDArray[np.uint8],
     offsets: NDArray[np.uint32],
-    reader: IElementReader,
+    reader: IReader,
 ) -> Any: ...
 def read_bes_raw(
     data: NDArray[np.uint32],
